+++ src/main/java/com/cavanosa/virtual/repository/TioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TioRepository extends JpaRepository<Tio, Integer> {
+public interface TioRepository extends JpaRepository<Tio, Long> {
     Optional<Tio> findByNombre(String nombre);
     Optional<Tio> findByEmail(String email);
     boolean existsByNombre(String nombre);
