+++ src/main/java/com/cavanosa/virtual/service/TioService.java
@@ -8,13 +8,13 @@ import java.util.Optional;
 public interface TioService {
 
     public List<Tio> findAll();
-    public Optional<Tio> getOneById(int id);
-    public Tio getFindById(int id);
+    public Optional<Tio> getOneById(Long id);
+    public Tio getFindById(Long id);
     public Optional<Tio> getOneByNombre(String nombre);
     public Optional<Tio> getOneByEmail(String email);
     public boolean save(Tio tio);
-    public void delete(int id);
-    public boolean existsById(int id);
+    public void delete(Long id);
+    public boolean existsById(Long id);
     public boolean existsByNombre(String nombre);
     public boolean exixtsByEmail(String email);
 }
